More of the same...

+++ util.c
@@ -296,17 +296,17 @@ checkexit_ext(int *rc, pid_t pid, enum comp comp)
 
 	if ( ! WIFEXITED(c) && WIFSIGNALED(c)) {
 		cp = strsignal(WTERMSIG(c));
-		warnx("signal: %s(%u): %s", comps[comp], pid, cp);
+		warnx("signal: %s(%lld): %s", comps[comp], (long long int) pid, cp);
 		return (0);
 	} else if ( ! WIFEXITED(c)) {
-		warnx("did not exit: %s(%u)", comps[comp], pid);
+		warnx("did not exit: %s(%lld)", comps[comp], (long long int) pid);
 		return (0);
 	}
 
 	/* Now check extended status. */
 
 	if (EXIT_SUCCESS != (*rc = WEXITSTATUS(c)) && 2 != *rc) {
-		dodbg("bad exit: %s(%u): %d", comps[comp], pid, *rc);
+		dodbg("bad exit: %s(%lld): %d", comps[comp], (long long int) pid, *rc);
 		return (0);
 	}
 	return (1);
